ref(post-process-forwarder): Completely remove post process worker
The post process worker was removed in https://github.com/getsentry/sentry/pull/44631.
as it has been replaced with an Arroyo implementation. This follow up removes the module
entirely and deletes some unused test fixtures.
