--- src/sentry/eventstream/kafka/postprocessworker.py
@@ -1,51 +1,0 @@
-from __future__ import annotations
-
-import logging
-import random
-from contextlib import contextmanager
-from enum import Enum
-from typing import Any, Generator, Mapping, Optional
-
-from sentry import options
-from sentry.eventstream.kafka.protocol import (
-    get_task_kwargs_for_message,
-    get_task_kwargs_for_message_from_headers,
-)
-from sentry.utils import metrics
-
-logger = logging.getLogger(__name__)
-
-Message = Any
-_DURATION_METRIC = "eventstream.duration"
-
-
-class PostProcessForwarderType(str, Enum):
-    ERRORS = "errors"
-    TRANSACTIONS = "transactions"
-    ISSUE_PLATFORM = "issue_platform"
-
-
-@contextmanager
-def _sampled_eventstream_timer(instance: str) -> Generator[None, None, None]:
-    record_metric = random.random() < 0.1
-    if record_metric is True:
-        with metrics.timer(_DURATION_METRIC, instance=instance):
-            yield
-    else:
-        yield
-
-
-def _get_task_kwargs(message: Message) -> Optional[Mapping[str, Any]]:
-    use_kafka_headers = options.get("post-process-forwarder:kafka-headers")
-
-    if use_kafka_headers:
-        try:
-            with _sampled_eventstream_timer(instance="get_task_kwargs_for_message_from_headers"):
-                return get_task_kwargs_for_message_from_headers(message.headers())
-        except Exception as error:
-            logger.error("Could not forward message: %s", error, exc_info=True)
-            with metrics.timer(_DURATION_METRIC, instance="get_task_kwargs_for_message"):
-                return get_task_kwargs_for_message(message.value())
-    else:
-        with metrics.timer(_DURATION_METRIC, instance="get_task_kwargs_for_message"):
-            return get_task_kwargs_for_message(message.value())
